fence_ibmz: add option --load-on-activate
Add support for the "Load during activation" option for IBM Z. This
option let's a system administrator specify a load device from which the
LPAR is loaded automatically during activation.

If the fence agent is invoked with the --load-on-activate command line
argument, the API request for an explicit load operation is skipped,
ultimately relying on the configuration of the "Load during activation"
option in the activation profile.

Signed-off-by: Bjoern Walk <bwalk@linux.ibm.com>
Signed-off-by: Bodo Brand <bodo.brand1@ibm.com>